Refactor: generateMeshResource Entity extension 함수로 변경
